ci: gate clang-tidy on changed lines and clear the backlog (#281)
Only the nightly runs clang-tidy, so a violation lands with nothing to
object and is found a day later. Ninja stops at the first failure, so each
night reported one finding and hid the rest.

Adds a lane over the changed lines, gated through CI OK. It reads the
compile database rather than the opt-in target list, so it also covers the
directories that never opted in - 539 translation units against 31 targets.
It analyses only files the database has a command for: a changed header has
none, and clang-tidy then falls back to a bare invocation where every symbol
in it reads as unprovided.

+++ libs/kernel/src/schedule_cycles.h
@@ -27,7 +27,7 @@ namespace sen::kernel
 
   // Divide before adding: the sum would overflow for a clock that jumps by decades, and an
   // overflowed count is negative, which would walk the schedule backwards.
-  return amount.count() / period.count() + (amount.count() % period.count() != 0);
+  return amount.count() / period.count() + static_cast<int64_t>(amount.count() % period.count() != 0);
 }
 
 }  // namespace sen::kernel
